Reorganized base files into base directory and some cleanup
