+++ lib/src/base/copy_path.dart
