+++ lib/src/context_runner.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:process/process.dart';
-import 'package:sylph/src/concurrent_jobs.dart';
 import 'package:tool_base/tool_base.dart';
 
 Future<T> runInContext<T>(
@@ -13,9 +12,7 @@ Future<T> runInContext<T>(
     body: runner,
     overrides: overrides,
     fallbacks: <Type, Generator>{
-//      AndroidSdk: AndroidSdk.locateAndroidSdk,
       BotDetector: () => const BotDetector(),
-      ConcurrentJobs: () => ConcurrentJobs(),
       Config: () => Config(),
       Logger: () => platform.isWindows ? WindowsStdoutLogger() : StdoutLogger(),
       OperatingSystemUtils: () => OperatingSystemUtils(),
