+++ lib/sylph.dart
@@ -1,5 +1,6 @@
 library sylph;
 
-export 'src/devices.dart';
+export 'src/base/devices.dart';
 export 'src/context_runner.dart';
+export 'src/device_farm.dart';
 export 'src/sylph_run.dart';
