+++ test/base/concurrent_jobs_test.dart
@@ -1,13 +1,10 @@
 import 'dart:async';
 
-import 'package:mockito/mockito.dart';
-import 'package:sylph/src/concurrent_jobs.dart';
+import 'package:sylph/src/base/concurrent_jobs.dart';
 import 'package:sylph/src/context_runner.dart';
 import 'package:test/test.dart';
 import 'package:tool_base/tool_base.dart';
 import 'package:tool_base_test/tool_base_test.dart';
-
-class MockConcurrentJobs extends Mock implements ConcurrentJobs {}
 
 main() {
   group('concurrent jobs', () {
@@ -17,7 +14,7 @@ main() {
           {'n': 10, 'verbose': true},
           {'n': 20, 'verbose': false}
         ];
-        List results = await concurrentJobs.runJobs(squareInContext, jobArgs);
+        List results = await runJobs(squareInContext, jobArgs);
         for (int i = 0; i < results.length; i++) {
           expect(results[i], await squareInContext(jobArgs[i]));
         }
@@ -26,7 +23,7 @@ main() {
       });
 
       testUsingContext('run sylph', () async {}, overrides: <Type, Generator>{
-        ConcurrentJobs: () => MockConcurrentJobs,
+//        ConcurrentJobs: () => MockConcurrentJobs,
 //        Logger: () => VerboseLogger(StdoutLogger()),
       });
     });
@@ -37,7 +34,7 @@ main() {
           {'n': 10},
           {'n': 20}
         ];
-        List results = await ConcurrentJobs().runJobs(square, jobArgs);
+        List results = await runJobs(square, jobArgs);
         for (int i = 0; i < results.length; i++) {
           expect(results[i], await square(jobArgs[i]));
         }
